# PERF: optimize for contiguous 2D nanmin/nanmax, speedups of up to 12x